backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Werkzeug==3.0.4
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe==10.5.0
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
PyPDF2==3.0.1
APScheduler==3.10.4
